Hide old versions of bamtools_split
When this tool was added to a suite, the tool got a new ID (without an
accompanying version bump or name change), which caused it to appear
twice under identical name, description and version in the same section
of the .eu tool panel.

The change here keeps only the latest version (id="bamtools_split")
visible and hides all versions with id="bamSplit".

See: galaxyproject/tools-iuc 864472b0